backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.4,<5
django-environ>=0.11,<1
dj-database-url>=2.2,<3
psycopg2-binary>=2.9,<3
gunicorn>=22.0,<23
requests>=2.32,<3
Pillow>=10.4,<11
django-filter>=24.3,<25